1.使用jaudiotagger增加一些元数据提取 2.增加lrc歌词解析 3.增加支持本地目录下专辑图片解析 4.增加播放信息页面

+++ media_query/src/main/java/com/hjkl/query/AlbumArtQuery.kt
@@ -14,7 +14,7 @@ class AlbumArtQuery : IAlbumArtQuery {
         private const val TAG = "AlbumArtQuery"
     }
 
-    override fun getAlbumArtBitmap(songId: Int, albumId: Int): Bitmap? {
+    override fun getAlbumArtBitmap(songId: Long, albumId: Int): Bitmap? {
         if (songId < 0 && albumId < 0) {
             return null
         }
