+++ media_query/src/main/java/com/hjkl/query/IAlbumArtQuery.kt
@@ -4,5 +4,5 @@ import android.graphics.Bitmap
 
 @Deprecated("有些机型获取不到专辑图片")
 interface IAlbumArtQuery {
-    fun getAlbumArtBitmap(songId: Int, albumId: Int): Bitmap?
+    fun getAlbumArtBitmap(songId: Long, albumId: Int): Bitmap?
 }
